+++ js/app.js
@@ -1,11 +1,72 @@
 // App Interface (MVC).
 
-// import necessary classes.
+/*
+The model class can be imported for a type driven UI, but this is optional.
+The datastore does not use model properties for persistence.
+All items are added to an item collection.
+The item collection is persisted as a JSON array, regardless of the persistence type selected.
+*/
 import { Book } from "./book.js";
-import { BookStore } from "./bookStore.js";
 
-// create a book collection manager.
-const bookDepot = new BookStore();
+/*
+The ItemStore module is required to manage the collection.
+The DatabaseSettings module is required to override the default persistence settings.
+See the example implementations below.
+*/
+import { ItemStore, DatabaseSettings } from "./itemStore.js";
+
+/*
+Persistence types is an enum of the browser supported persistence methods.
+This module can be imported for strongly typed declarations but his is optional.
+The persistence types can be passed directly as strings.
+*/
+import { PersistenceTypes } from "./dataService.js";
+
+/***************************************************************************************
+Use this declaration to create a item store to manage the collection using default settings.
+The collection will be in memory only and deleted upon page refresh or when the browser window is closed.
+****************************************************************************************/
+
+//const bookDepot = new ItemStore();
+
+/*************************************************************************************
+Use this declaration to create a item store to manage the collection without persistence.
+When the usePersistenceStore parameter is set to false,
+the collection will be in memory only and deleted upon page refresh or when the browser window is closed.
+*************************************************************************************/
+
+//const bookDepot = new ItemStore("MyBookStore", false);
+
+/*************************************************************************************
+Use this declaration to create an item store to manage the collection with default persistence settings.
+This implementation will use cookie storage by default.
+The collection will be stored with the browser's cookie until the cookie expires or is deleted.
+**************************************************************************************/
+
+const bookDepot = new ItemStore("MyBookStore", true);
+
+/**************************************************************************************
+An instace of the DatabaseSettings module is required to override any of the default settings.
+Settings such as the database name, database version, table name, as well as key field can be overridden.
+This is also where the persistence type can be changed from cookie storage to another option.
+This code is ignored when the ItemStore.usePersistence property is set to false.
+***************************************************************************************/
+
+// Passing in values only.
+//bookDepot.initializeDataStore("MyBookStore", 1, "Books", "id", "cookie");
+
+// Using a settings object.
+var settings = new DatabaseSettings();
+
+settings.databaseName = "MyBookStore";
+settings.databaseVersion = 2;
+settings.objectStoreName = "Books";
+settings.keyPathField = "id";
+settings.persistenceType = PersistenceTypes.LocalStorage;
+
+bookDepot.initializeDataStore(settings);
+
+/**************************************************************************************/
 
 // new page title
 const pageTitle = "Javascript - Bookshelf";
@@ -31,6 +92,7 @@ const markUp =
   '<label for="new-book-numberOfPages" style="margin-left:8px">Page count:</label><input type="number" id="new-book-numberOfPages"  style="display:inline-block; margin-left:8px" />' +
   '<button id="new-book-form-button" type="submit"  style="display:inline-block; margin-left:8px">Save</button>' +
   '<button id="clear-form" type="reset"  style="display:inline-block; margin-left:8px">Clear</button>' +
+  '<button id="refresh-data" type="refresh"  style="display:inline-block; margin-left:8px">Refresh</button>' +
   "</form>" +
   "</fieldset>" +
   "</div>" +
@@ -69,12 +131,23 @@ const markUp =
       event.preventDefault();
     });
 
+  document
+    .getElementById("refresh-data")
+    .addEventListener("click", function (event) {
+      reloadData();
+      event.preventDefault();
+    });
+
   // seed the collection and update the display
   seedTheRepository();
 
   // clear the form
   resetForm();
 })();
+
+function reloadData() {
+  seedTheRepository();
+}
 
 /*
     Clear all form values.