+++ changes.md
@@ -6,3 +6,9 @@
 - Improved code modularity.
 - Added Changes.md to track changes.
 - Updated README.md.
+
+#### June 20, 2024
+
+- Moved to timestamp for ID field to avoid collisions.
+- Refactored dataService.js
+- Implemented localStorage and sessionStorage services.